Inline emitAnnotations
